Program in branches reworked for partial optimization
This is the program already committed in branches (please, use them instead of commiting your changes directlly on main; we could discuss and modify them before merging them into the main). 
Anyway, the program is the same except that uses thread for initial import and it have a simple GUI for resizing the shape of the selected area. Also i cancelled the old CNN in order to give space for the better one you wrote (So it includes only the SVC-RFE).